and fix yet another picky warning
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@17063 a1c6a512-1295-4272-9138-f99709370657

+++ tools/convbdf.c
@@ -738,7 +738,7 @@ int rotleft(unsigned char *dst, /* output buffer */
 
     if(((height + 7) / 8) * width > dstlen) {
         fprintf(stderr, "%s:%d %d x %d overflows %d bytes buffer, needs %d\n",
-                __FILE__, __LINE__, width, height, dstlen,
+                __FILE__, __LINE__, width, height, (int)dstlen,
                 ((height + 7) / 8) * width );
         return 0;
     }
